Move _mysql and _mysql_exceptions into MySQLdb/ (#293)
