+++ MySQLdb/_mysql.c
@@ -2746,7 +2746,7 @@ init_mysql(void)
                    (PyObject *)&_mysql_ResultObject_Type))
         goto error;
     Py_INCREF(&_mysql_ResultObject_Type);
-    if (!(emod = PyImport_ImportModule("_mysql_exceptions"))) {
+    if (!(emod = PyImport_ImportModule("MySQLdb._mysql_exceptions"))) {
         PyErr_Print();
         goto error;
     }
