+++ MySQLdb/_mysql_exceptions.py
