+++ MySQLdb/cursors.py
@@ -8,8 +8,8 @@ from functools import partial
 import re
 import sys
 
-from MySQLdb.compat import unicode
-from _mysql_exceptions import (
+from .compat import unicode
+from ._mysql_exceptions import (
     Warning, Error, InterfaceError, DataError,
     DatabaseError, OperationalError, IntegrityError, InternalError,
     NotSupportedError, ProgrammingError)
@@ -55,9 +55,11 @@ class BaseCursor(object):
     #: Default value of max_allowed_packet is 1048576.
     max_stmt_length = 64*1024
 
-    from _mysql_exceptions import MySQLError, Warning, Error, InterfaceError, \
-         DatabaseError, DataError, OperationalError, IntegrityError, \
-         InternalError, ProgrammingError, NotSupportedError
+    from ._mysql_exceptions import (
+        MySQLError, Warning, Error, InterfaceError,
+        DatabaseError, DataError, OperationalError, IntegrityError,
+        InternalError, ProgrammingError, NotSupportedError,
+    )
 
     _defer_warnings = False
     connection = None
