+++ metadata.cfg
@@ -31,7 +31,7 @@ classifiers:
         Topic :: Database
         Topic :: Database :: Database Engines/Servers
 py_modules:
-        _mysql_exceptions
+        MySQLdb._mysql_exceptions
         MySQLdb.compat
         MySQLdb.connections
         MySQLdb.converters
