+++ setup.py
@@ -14,7 +14,9 @@ with io.open('README.md', encoding='utf-8') as f:
     readme = f.read()
 
 metadata, options = get_config()
-metadata['ext_modules'] = [setuptools.Extension(sources=['_mysql.c'], **options)]
+metadata['ext_modules'] = [
+    setuptools.Extension("MySQLdb._mysql", sources=['MySQLdb/_mysql.c'], **options)
+]
 metadata['long_description'] = readme
 metadata['long_description_content_type'] = "text/markdown"
 setuptools.setup(**metadata)
