+++ setup_windows.py
@@ -35,7 +35,6 @@ def get_config():
     create_release_file(metadata)
     del metadata['version_info']
     ext_options = dict(
-        name = "_mysql",
         library_dirs = library_dirs,
         libraries = libraries,
         extra_compile_args = extra_compile_args,
@@ -43,7 +42,7 @@ def get_config():
         include_dirs = include_dirs,
         extra_objects = extra_objects,
         define_macros = define_macros,
-        )
+    )
     return metadata, ext_options
 
 if __name__ == "__main__":
