+++ tests/test__mysql.py
@@ -1,5 +1,5 @@
 import pytest
-import _mysql
+from MySQLdb import _mysql
 
 
 def test_result_type():
